backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
SQLAlchemy==2.0.32
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
websockets==12.0
slowapi==0.1.9
python-multipart==0.0.9
itsdangerous==2.2.0
starlette==0.37.2


